Bugfixes + neue Qry
git-svn-id: http://isdb1.offis.uni-oldenburg.de/repos/odysseus/trunk@8376 aa0bda67-78ed-0310-a382-c5338b0f89ba

Former-commit-id: 6eda9ed26aa3828a4af4bb40af6c5d382eba9b11

+++ sensor/streamcars/scars.base/src/de/uniol/inf/is/odysseus/scars/base/JDVEAccessMVPO.java
@@ -213,8 +213,8 @@ class JDVEData<M extends IProbability> {
 	public MVRelationalTuple<M> parseList(SDFAttribute schema, ByteBuffer bb) {
 		// int count = bb.getInt(); // TODO: hier Länge aus Buffer einlesen
 		// System.out.println(count);
-		int count = bb.getInt();
-		MVRelationalTuple<M> recordTuple = new MVRelationalTuple<M>(count);
+		long count = bb.getLong();
+		MVRelationalTuple<M> recordTuple = new MVRelationalTuple<M>((int)count);
 
 		for (int i = 0; i < count; i++) {
 			Object obj = parseNext(schema.getSubattribute(0), bb);
